+++ ICD.Profound.ConnectPRO/Rooms/ConnectProRoom.cs
@@ -17,6 +17,7 @@ using ICD.Connect.Displays.Devices;
 using ICD.Connect.Panels;
 using ICD.Connect.Panels.Server.Osd;
 using ICD.Connect.Partitioning.Rooms;
+using ICD.Connect.Partitioning.VolumePoints;
 using ICD.Connect.Routing.Endpoints.Destinations;
 using ICD.Connect.Settings.Core;
 using ICD.Profound.ConnectPRO.Routing;
@@ -32,7 +33,6 @@ namespace ICD.Profound.ConnectPRO.Rooms
 
 		private bool m_IsInMeeting;
 		private DialingPlanInfo m_DialingPlan;
-		private VolumePoint m_VolumePoint;
 
 		#region Properties
 
@@ -71,9 +71,11 @@ namespace ICD.Profound.ConnectPRO.Rooms
 		[CanBeNull]
 		public IVolumeDeviceControl GetVolumeControl()
 		{
-			return m_VolumePoint == null
+			IVolumePoint volumePoint = Originators.GetInstance<IVolumePoint>();
+
+			return volumePoint == null
 				       ? null
-				       : Core.GetControl<IVolumeDeviceControl>(m_VolumePoint.DeviceId, m_VolumePoint.ControlId);
+					   : Core.GetControl<IVolumeDeviceControl>(volumePoint.DeviceId, volumePoint.ControlId);
 		}
 
 		/// <summary>
@@ -166,7 +168,6 @@ namespace ICD.Profound.ConnectPRO.Rooms
 			base.CopySettingsFinal(settings);
 
 			settings.DialingPlan = m_DialingPlan;
-			settings.VolumePoint = m_VolumePoint;
 		}
 
 		/// <summary>
@@ -177,8 +178,6 @@ namespace ICD.Profound.ConnectPRO.Rooms
 			base.ClearSettingsFinal();
 
 			m_DialingPlan = default(DialingPlanInfo);
-
-			m_VolumePoint = null;
 
 			m_ConferenceManager.ClearDialingProviders();
 			m_ConferenceManager.Favorites = null;
@@ -197,21 +196,9 @@ namespace ICD.Profound.ConnectPRO.Rooms
 			// Dialing plan
 			SetDialingPlan(settings.DialingPlan, factory);
 
-			// Volume Point
-			SetVolumePoint(settings.VolumePoint);
-
 			// Favorites
 			string path = PathUtils.GetProgramConfigPath("favorites");
 			m_ConferenceManager.Favorites = new SqLiteFavorites(path);
-		}
-
-		/// <summary>
-		/// Sets the volume point from the settings.
-		/// </summary>
-		/// <param name="volumePoint"></param>
-		private void SetVolumePoint(VolumePoint volumePoint)
-		{
-			m_VolumePoint = volumePoint;
 		}
 
 		/// <summary>